+++ CHANGES.md
@@ -3,6 +3,8 @@ Changes
 
 0.1.3 (unreleased)
 
+- Fix output of Windows paths in magik-lint
+
 
 0.1.2 (2019-07-06)